Getting rid of white lines
I looked at the code in histogram.go, and determined it is complicated
and I'm best off observing the step size it chooses and using that.

Procedures for this:
1) run a heatmap with the pixels I want
2) open browser network tab
3) rerun query
4) find ws messages
5) copy the content of the last message, which starts with 'series'
6) look at the step size in there
7) change the code to squash the picture so that the step size represents 1 pixel

this involved changing the height of the image, too. ya can just change the height
in the heatmap generation, the image can be shorter whatevs